refactor(qt): split modules
