+++ src/ui/qt/widget.rs
@@ -5,68 +5,6 @@ pub(crate) use ffi::*;
 
 use crate::{Point, Rect, Size};
 
-#[repr(C)]
-#[doc(hidden)]
-pub struct QRect {
-    pub x1: i32,
-    pub y1: i32,
-    pub x2: i32,
-    pub y2: i32,
-}
-
-unsafe impl ExternType for QRect {
-    type Id = type_id!("QRect");
-    type Kind = cxx::kind::Trivial;
-}
-
-#[cxx::bridge]
-mod ffi {
-    unsafe extern "C++" {
-        include!("winio/src/ui/qt/widget.hpp");
-
-        fn is_dark() -> bool;
-
-        type QWidget;
-        type QRect = super::QRect;
-
-        fn x(self: &QWidget) -> i32;
-        fn y(self: &QWidget) -> i32;
-        #[cxx_name = "move"]
-        fn move_(self: Pin<&mut QWidget>, x: i32, y: i32);
-        fn width(self: &QWidget) -> i32;
-        fn height(self: &QWidget) -> i32;
-        fn resize(self: Pin<&mut QWidget>, w: i32, h: i32);
-        fn geometry(self: &QWidget) -> &QRect;
-
-        fn widget_get_title(w: &QWidget) -> String;
-        fn widget_set_title(w: Pin<&mut QWidget>, s: &str);
-
-        fn new_main_window() -> UniquePtr<QWidget>;
-        unsafe fn main_window_register_resize_event(
-            w: Pin<&mut QWidget>,
-            callback: unsafe fn(*const u8, i32, i32),
-            data: *const u8,
-        );
-        unsafe fn main_window_register_move_event(
-            w: Pin<&mut QWidget>,
-            callback: unsafe fn(*const u8, i32, i32),
-            data: *const u8,
-        );
-        unsafe fn main_window_register_close_event(
-            w: Pin<&mut QWidget>,
-            callback: unsafe fn(*const u8) -> bool,
-            data: *const u8,
-        );
-
-        fn new_push_button(parent: Pin<&mut QWidget>) -> UniquePtr<QWidget>;
-        unsafe fn push_button_connect_clicked(
-            w: Pin<&mut QWidget>,
-            callback: unsafe fn(*const u8),
-            data: *const u8,
-        );
-    }
-}
-
 pub struct Widget {
     widget: RefCell<UniquePtr<QWidget>>,
 }
@@ -119,5 +57,47 @@ impl Widget {
     pub fn set_text(&self, s: &str) {
         let mut widget = self.widget.borrow_mut();
         widget_set_title(widget.pin_mut(), s);
+    }
+}
+
+#[repr(C)]
+#[doc(hidden)]
+pub struct QRect {
+    pub x1: i32,
+    pub y1: i32,
+    pub x2: i32,
+    pub y2: i32,
+}
+
+unsafe impl ExternType for QRect {
+    type Id = type_id!("QRect");
+    type Kind = cxx::kind::Trivial;
+}
+
+#[cxx::bridge]
+mod ffi {
+    unsafe extern "C++" {
+        include!("winio/src/ui/qt/widget.hpp");
+        include!("winio/src/ui/qt/window.hpp");
+
+        fn is_dark() -> bool;
+
+        type QWidget;
+        type QRect = super::QRect;
+
+        #[allow(dead_code)]
+        fn new_main_window() -> UniquePtr<QWidget>;
+
+        fn x(self: &QWidget) -> i32;
+        fn y(self: &QWidget) -> i32;
+        #[cxx_name = "move"]
+        fn move_(self: Pin<&mut QWidget>, x: i32, y: i32);
+        fn width(self: &QWidget) -> i32;
+        fn height(self: &QWidget) -> i32;
+        fn resize(self: Pin<&mut QWidget>, w: i32, h: i32);
+        fn geometry(self: &QWidget) -> &QRect;
+
+        fn widget_get_title(w: &QWidget) -> String;
+        fn widget_set_title(w: Pin<&mut QWidget>, s: &str);
     }
 }
